simplifies code by embedding pre state initialisation in factory
